+++ src/DSharp.Compiler.Tests/DSharp.Compiler.Tests.csproj
@@ -71,8 +71,8 @@
     <Compile Remove="Source\Member\Overloads\Code.cs" />
     <Compile Remove="Source\Member\Properties\Code.cs" />
     <Compile Remove="Source\Member\StaticConstructors\Code.cs" />
+    <Compile Remove="Source\Mscorlib\Collections\Code.cs" />
     <Compile Remove="Source\Mscorlib\DateTime\Code.cs" />
-    <Compile Remove="Source\Mscorlib\CustomCollections\Code.cs" />
     <Compile Remove="Source\Statement\Exceptions\Code.cs" />
     <Compile Remove="Source\Statement\Expression\Code.cs" />
     <Compile Remove="Source\Statement\Foreach\Code.cs" />
@@ -185,7 +185,7 @@
     <None Remove="Source\Member\Properties\Baseline.txt" />
     <None Remove="Source\Member\StaticConstructors\Baseline.txt" />
     <None Remove="Source\Mscorlib\DateTime\Baseline.txt" />
-    <None Remove="Source\Mscorlib\CustomCollections\Baseline.txt" />
+    <None Remove="Source\Mscorlib\Collections\Baseline.txt" />
     <None Remove="Source\Statement\Exceptions\Baseline.txt" />
     <None Remove="Source\Statement\Expression\Baseline.txt" />
     <None Remove="Source\Statement\Foreach\Baseline.txt" />
@@ -579,16 +579,16 @@
     <Content Include="Source\Member\StaticConstructors\Code.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Content Include="Source\Mscorlib\Collections\Code.cs">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <Content Include="Source\Mscorlib\DateTime\Baseline.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Source\Mscorlib\DateTime\Code.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="Source\Mscorlib\CustomCollections\Baseline.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="Source\Mscorlib\CustomCollections\Code.cs">
+    <Content Include="Source\Mscorlib\Collections\Baseline.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Source\Statement\Exceptions\Baseline.txt">
