+++ src/DSharp.Compiler.Tests/Source/CompilerTests.json
@@ -8,7 +8,7 @@
         "ComparisonFile": "Baseline.txt"
       },
       {
-        "Name": "ICollection",
+        "Name": "Collections",
         "SourceFiles": [ "Code.cs" ],
         "ComparisonFile": "Baseline.txt"
       }
