+++ src/DSharp.Compiler.Tests/Source/Mscorlib/Collections/Baseline.txt
@@ -1,0 +1,140 @@
+﻿"use strict";
+
+define('test', ['ss'], function(ss) {
+  var $global = this;
+  // MscorlibTests.MyCollection
+
+  function MyCollection() {
+  }
+  var MyCollection$ = {
+    $get_length: function() {
+      return 2;
+    },
+    $get_isReadOnly: function() {
+      return false;
+    },
+    push: function(item) {
+    },
+    clear: function() {
+    },
+    contains: function(item) {
+      return false;
+    },
+    copyTo: function(array, arrayIndex) {
+    },
+    getEnumerator: function() {
+      return null;
+    },
+    remove: function(item) {
+      return false;
+    }
+  };
+
+
+  // MscorlibTests.X
+
+  function X() {
+  }
+  X.testList = function() {
+    var c = [];
+    c.push(12);
+    c.push(c);
+    ss.clear(c);
+    ss.contains(c, 12);
+    ss.removeItem(c, 12);
+    ss.removeAt(c, 1);
+    var e = ss.enumerate(c);
+    var i = c.indexOf(12);
+    c.forEach(function(n) {
+      ++n;
+    });
+    var cc = c.length;
+  };
+  X.testIList_T = function() {
+    var c = [];
+    var n = c[0];
+    c[0] = c[1];
+    c.push(12);
+    ss.clear(c);
+    ss.contains(c, 12);
+    ss.removeItem(c, 12);
+    ss.removeAt(c, 1);
+    var e = ss.enumerate(c);
+    var i = c.indexOf(12);
+    var cc = c.length;
+  };
+  X.testIList = function() {
+    var c = [];
+    var n = c[0];
+    c[0] = c[1];
+    c.push(12);
+    ss.clear(c);
+    ss.contains(c, 12);
+    ss.removeItem(c, 12);
+    ss.removeAt(c, 1);
+    var e = ss.enumerate(c);
+    var i = c.indexOf(12);
+    var cc = c.length;
+  };
+  X.testICollection_T = function() {
+    var c = [];
+    c.push(12);
+    ss.clear(c);
+    ss.contains(c, 12);
+    ss.removeItem(c, 12);
+    var e = ss.enumerate(c);
+    var cc = c.length;
+  };
+  X.testICollection = function() {
+    var c = [];
+    var e = ss.enumerate(c);
+    var cc = c.length;
+  };
+  X.testIReadOnlyList = function() {
+    var c = [];
+    var n = c[0];
+    var e = ss.enumerate(c);
+    var cc = c.length;
+  };
+  X.testArray = function() {
+    var c = new Array(4);
+    var cl = c.length;
+    var e = ss.enumerate(c);
+  };
+  X.testIReadOnlyCollection = function() {
+    var c = [];
+    var e = ss.enumerate(c);
+    var cc = c.length;
+  };
+  X.testIEnumerable_T = function() {
+    var c = [];
+    var e = ss.enumerate(c);
+  };
+  X.testIEnumerable = function() {
+    var c = [];
+    var e = ss.enumerate(c);
+  };
+  X.testCustomList = function() {
+    var c = new MyCollection();
+    var n = c.get_item(0);
+    c.set_item(0, c.get_item(1));
+    c.push(12);
+    ss.clear(c);
+    ss.contains(c, 12);
+    ss.removeItem(c, 12);
+    ss.removeAt(c, 1);
+    var e = ss.enumerate(c);
+    var i = c.indexOf(12);
+    var cc = c.length;
+  };
+
+
+  var $exports = ss.module('test', null,
+    {
+      MyCollection: ss.defineClass(MyCollection, MyCollection$, [], null, [ss.ICollection]),
+      X: ss.defineClass(X, null, [], null)
+    });
+
+
+  return $exports;
+});
