+++ src/DSharp.Compiler.Tests/Source/Mscorlib/Collections/Code.cs
@@ -1,0 +1,193 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+
+[assembly: ScriptAssembly("test")]
+
+namespace MscorlibTests
+{
+    public class MyCollection : IList<int>
+    {
+        public int Count { get { return 2; } }
+
+        public bool IsReadOnly { get { return false; } }
+
+        public void Add(int item)
+        {
+        }
+
+        public void Clear()
+        {
+        }
+
+        public bool Contains(int item)
+        {
+            return false;
+        }
+
+        public void CopyTo(int[] array, int arrayIndex)
+        {
+        }
+
+        public IEnumerator<int> GetEnumerator()
+        {
+            return null;
+        }
+
+        public bool Remove(int item)
+        {
+            return false;
+        }
+
+        public int this[int index] { get { return 1; } set { } }
+
+        public int IndexOf(int item) { return 1; }
+
+        public void Insert(int index, int item) { }
+
+        public void RemoveAt(int index) { }
+
+        [ScriptIgnore]
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            return null;
+        }
+    }
+
+    public static class X
+    {
+        public static void TestList()
+        {
+            List<int> c = new List<int>();
+            c.Add(12);
+            c.AddRange(c); // remove method or wrap in spread operator
+            c.Clear();
+            c.Contains(12);
+            c.Remove(12);
+            c.RemoveAt(1);
+            IEnumerator<int> e = c.GetEnumerator();
+            int i = c.IndexOf(12);
+            c.ForEach(delegate (int n) { ++n; });
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+            //c.Insert(1, 22);
+            //c.CopyTo(null);
+        }
+
+        public static void TestIList_T()
+        {
+            IList<int> c = new List<int>();
+            int n = c[0];
+            c[0] = c[1];
+            c.Add(12);
+            c.Clear();
+            c.Contains(12);
+            c.Remove(12);
+            c.RemoveAt(1);
+            IEnumerator<int> e = c.GetEnumerator();
+            int i = c.IndexOf(12);
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+            //c.Insert(1, 22);
+            //c.CopyTo(null, 0);
+        }
+
+        public static void TestIList()
+        {
+            IList c = new List<int>();
+            object n = c[0];
+            c[0] = c[1];
+            c.Add(12);
+            c.Clear();
+            c.Contains(12);
+            c.Remove(12);
+            c.RemoveAt(1);
+            IEnumerator e = c.GetEnumerator();
+            int i = c.IndexOf(12);
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+            //c.Insert(1, 22);
+            //c.CopyTo(null, 0);
+        }
+
+        public static void TestICollection_T()
+        {
+            ICollection<int> c = new List<int>();
+            c.Add(12);
+            c.Clear();
+            c.Contains(12);
+            c.Remove(12);
+            IEnumerator<int> e = c.GetEnumerator();
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+            //c.CopyTo(null, 0);
+        }
+
+        public static void TestICollection()
+        {
+            ICollection c = new List<int>();
+            IEnumerator e = c.GetEnumerator();
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+            //c.CopyTo(null, 0);
+        }
+
+        public static void TestIReadOnlyList()
+        {
+            IReadOnlyList<int> c = new List<int>().AsReadOnly();
+            object n = c[0];
+            IEnumerator e = c.GetEnumerator();
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+        }
+
+        public static void TestArray()
+        {
+            Array c = new int[4];
+            int cl = c.Length;
+            IEnumerator e = c.GetEnumerator();
+            //bool eq = c.Equals(c);
+            //c.CopyTo(null, 0);
+        }
+
+        public static void TestIReadOnlyCollection()
+        {
+            IReadOnlyCollection<int> c = new List<int>().AsReadOnly();
+            IEnumerator e = c.GetEnumerator();
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+        }
+
+        public static void TestIEnumerable_T()
+        {
+            IEnumerable<int> c = new List<int>();
+            IEnumerator<int> e = c.GetEnumerator();
+            //bool eq = c.Equals(c);
+        }
+
+        public static void TestIEnumerable()
+        {
+            IEnumerable c = new List<int>();
+            IEnumerator e = c.GetEnumerator();
+            //bool eq = c.Equals(c);
+        }
+
+        public static void TestCustomList()
+        {
+            MyCollection c = new MyCollection();
+            int n = c[0];
+            c[0] = c[1];
+            c.Add(12);
+            c.Clear();
+            c.Contains(12);
+            c.Remove(12);
+            c.RemoveAt(1);
+            IEnumerator<int> e = c.GetEnumerator();
+            int i = c.IndexOf(12);
+            int cc = c.Count;
+            //bool eq = c.Equals(c);
+            //c.Insert(1, 22);
+            //c.CopyTo(null, 0);
+        }
+    }
+}
