--- src/DSharp.Compiler.Tests/Source/Mscorlib/CustomCollections/Baseline.txt
@@ -1,1 +1,0 @@
-﻿
